Move LogoutUseCase in its own module
Now no longer needs to know of the world which was the original reason
that the implementation was in the app module
